chore: move files about review
